backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
aiomysql==0.2.0
cryptography
PyJWT==2.8.0
pandas==2.2.3
tensorflow==2.16.1
numpy==1.26.4
scikit-learn==1.5.2